Fix race condition with model deletion
- There is a case where the model can be loaded, but then a parallel
deletion job deletes the model from the DB before it can be deleted.
This results in a Sequel error saying that the delete did not result in any
row modification.
- This was surfacing on some environments for route mappings due to some
clients deleting the same route in parallel

[#154617293]

+++ app/jobs/runtime/model_deletion.rb
@@ -15,7 +15,14 @@ module VCAP::CloudController
 
           model = model_class.find(guid: guid)
           return if model.nil?
-          model.destroy
+
+          model.db.transaction do
+            if model.exists?
+              model.lock!
+
+              model.destroy
+            end
+          end
         end
 
         def job_name_in_configuration
